+++ fabric/src/main/kotlin/net/asch/asc/fabric/ui/screen/PosePresetScreen.kt
@@ -1,0 +1,24 @@
+package net.asch.asc.fabric.ui.screen
+
+import io.wispforest.owo.ui.base.BaseOwoScreen
+import io.wispforest.owo.ui.container.Containers
+import io.wispforest.owo.ui.container.FlowLayout
+import io.wispforest.owo.ui.core.HorizontalAlignment
+import io.wispforest.owo.ui.core.OwoUIAdapter
+import io.wispforest.owo.ui.core.Surface
+import io.wispforest.owo.ui.core.VerticalAlignment
+import net.asch.asc.fabric.ui.component.PresetPanel
+
+class PosePresetScreen : BaseOwoScreen<FlowLayout>() {
+    override fun createAdapter(): OwoUIAdapter<FlowLayout> {
+        return OwoUIAdapter.create(this, Containers::verticalFlow)
+    }
+
+    override fun build(rootComponent: FlowLayout) {
+        rootComponent.surface(Surface.VANILLA_TRANSLUCENT)
+        rootComponent.horizontalAlignment(HorizontalAlignment.CENTER)
+        rootComponent.verticalAlignment(VerticalAlignment.CENTER)
+
+        rootComponent.child(PresetPanel())
+    }
+}