+++ src/client/kotlin/com/asc/AscClient.kt
@@ -1,9 +1,0 @@
-package com.asc
-
-import net.fabricmc.api.ClientModInitializer
-
-class AscClient : ClientModInitializer {
-    override fun onInitializeClient() {
-        AscKeybinding.registerKeybindings()
-    }
-}