+++ CHANGELOG.md
@@ -1,5 +1,28 @@
-### Armor Statues Companion v1.2.1+1.20.1
+## [v1.3.0+1.20.1]
 
-* UI overall:
-  * Switch to left alignment
-  * Added colored buttons
+---
+
+### [`Armor Statues Datapack`](https://www.curseforge.com/minecraft/customization/armor-statues-datapack) support
+
+---
+
+- Added preset support
+- Added wands support
+- Added undo/redo/repeat support
+
+### API
+
+---
+
+- Multiplatform support via [`Architectury`](https://modrinth.com/mod/architectury-api) (only Fabric supported right now)
+- Replace `GitHub Actions` by [`Minotaur`](https://github.com/modrinth/minotaur) for deployment to [`Modrinth`](https://modrinth.com)
+
+### Changed
+
+---
+
+- New mod icon
+- (GUI) Switched to a "Toolbar + toolbox" design for a cleaner design
+- (GUI) Added a *Pose Preset* screen
+- (Singleplayer) GUI doesn't pause the game
+- (Singleplayer) Action triggers can now be used without closing te GUI